refactor #58: FAB 컴포넌트 분리
- Main 컴포넌트 복잡성을 줄이기 위한 컴포넌트 분리
