+++ fe/src/components/FAB.tsx
@@ -1,0 +1,30 @@
+import { COLOR_VARIANTS } from "@constants/colors";
+import { useFetch } from "hooks/useFetch";
+import { Button } from "./base/Button";
+import { PlusIcon } from "./icon/PlusIcon";
+
+export const FAB = ({ onColumnChanged }: { onColumnChanged: () => void }) => {
+  const { fetch: columnAddFetch } = useFetch({
+    url: "/api/columns",
+    method: "post",
+    body: {
+      columnName: "새로운 컬럼",
+    },
+  });
+
+  const onColumnAdd = async () => {
+    await columnAddFetch();
+    onColumnChanged();
+  };
+
+  return (
+    <Button
+      variant="blue"
+      pattern="FAB"
+      onClick={onColumnAdd}
+      css={{ position: "fixed", bottom: "30px", right: "30px" }}
+    >
+      <PlusIcon size={32} rgb={COLOR_VARIANTS.textWhiteDefault} />
+    </Button>
+  );
+};
